chore(s3-assets): use jest for tests (#8411)
No new tests or expectations added.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
