+++ packages/@aws-cdk/aws-s3-assets/.gitignore
@@ -15,3 +15,4 @@ nyc.config.js
 .LAST_PACKAGE
 *.snk
 !.eslintrc.js
+!jest.config.js
