+++ packages/@aws-cdk/aws-s3-assets/.npmignore
@@ -19,3 +19,4 @@ dist
 
 tsconfig.json
 .eslintrc.js
+jest.config.js
