+++ packages/@aws-cdk/aws-s3-assets/jest.config.js
@@ -1,0 +1,2 @@
+const baseConfig = require('../../../tools/cdk-build-tools/config/jest.config');
+module.exports = baseConfig;
