+++ packages/@aws-cdk/aws-s3-assets/package.json
@@ -45,6 +45,9 @@
     "build+test": "npm run build && npm test",
     "compat": "cdk-compat"
   },
+  "cdk-build": {
+    "jest": true
+  },
   "keywords": [
     "aws",
     "cdk",
@@ -60,16 +63,10 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@aws-cdk/assert": "0.0.0",
-    "@types/nodeunit": "^0.0.31",
-    "@types/sinon": "^9.0.4",
-    "aws-cdk": "0.0.0",
     "cdk-build-tools": "0.0.0",
     "cdk-integ-tools": "0.0.0",
-    "nodeunit": "^0.11.3",
     "pkglint": "0.0.0",
-    "sinon": "^9.0.2",
-    "@aws-cdk/cloud-assembly-schema": "0.0.0",
-    "ts-mock-imports": "^1.3.0"
+    "@aws-cdk/cloud-assembly-schema": "0.0.0"
   },
   "dependencies": {
     "@aws-cdk/assets": "0.0.0",
@@ -93,9 +90,6 @@
   },
   "stability": "experimental",
   "maturity": "experimental",
-  "nyc": {
-    "statements": 75
-  },
   "awslint": {
     "exclude": [
       "docs-public-apis:@aws-cdk/aws-s3-assets.AssetOptions",
